+++ src/routes/message.ts
@@ -1,14 +1,16 @@
 const router_message = require("express").Router();
-import { messageController } from "../controllers/messagecontroller"
-
+import {getMessages, getOneMessage} from "../controllers/messages/getmessage";
+import {createMessage} from "../controllers/messages/createmessage";
+import {deleteMessage} from "../controllers/messages/deletemessage";
+import {updateMessage} from "../controllers/messages/updatemessage";
 
 router_message.route("/:channelId/messages")
-  .get(messageController.getMessages)
-  .post(messageController.createMessage);
+  .get(getMessages)
+  .post(createMessage);
 
 router_message.route("/:channelId/messages/:messageId")
-  .get(messageController.getOneMessage)
-  .patch(messageController.updateMessage)
-  .delete(messageController.deleteMessage)
+  .get(getOneMessage)
+  .patch(updateMessage)
+  .delete(deleteMessage)
 
 module.exports = router_message;
